+++ src/main/java/org/example/sinara/repository/sql/PagamantoRepository.java
@@ -3,5 +3,5 @@ package org.example.sinara.repository.sql;
 import org.example.sinara.model.Pagamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagamantoRepository extends JpaRepository<Pagamento, Long> {
+public interface PagamantoRepository extends JpaRepository<Pagamento, Integer> {
 }
